+++ rastervision_core/rastervision/core/rv_pipeline/semantic_segmentation.py
@@ -1,18 +1,24 @@
 import logging
-from typing import List
+from typing import List, Sequence
 
 import numpy as np
 
-from rastervision.core.rv_pipeline.rv_pipeline import RVPipeline
 from rastervision.core.box import Box
+from rastervision.core.data import ClassConfig, Scene
+from rastervision.core.rv_pipeline.rv_pipeline import RVPipeline
+from rastervision.core.rv_pipeline.utils import (fill_no_data,
+                                                 nodata_below_threshold)
 from rastervision.core.rv_pipeline.semantic_segmentation_config import (
-    SemanticSegmentationWindowMethod)
+    SemanticSegmentationWindowMethod, SemanticSegmentationChipOptions)
 
 log = logging.getLogger(__name__)
 
 
-def get_train_windows(scene, class_config, chip_size,
-                      chip_options) -> List[Box]:
+def get_train_windows(scene: Scene,
+                      class_config: ClassConfig,
+                      chip_size: int,
+                      chip_options: SemanticSegmentationChipOptions,
+                      chip_nodata_threshold: float = 1.) -> List[Box]:
     """Get training windows covering a scene.
 
     Args:
@@ -26,21 +32,36 @@ def get_train_windows(scene, class_config, chip_size,
     extent = raster_source.get_extent()
     label_source = scene.ground_truth_label_source
 
-    def filter_windows(windows):
+    def filter_windows(windows: Sequence[Box]) -> List[Box]:
+        """Filter out chips that
+        (1) are outside the AOI
+        (2) only consist of null labels
+        (3) have NODATA proportion >= chip_nodata_threshold
+        """
+        total_windows = len(windows)
         if scene.aoi_polygons:
             windows = Box.filter_by_aoi(windows, scene.aoi_polygons)
+            log.info(f'AOI filtering: {len(windows)}/{total_windows} '
+                     'chips accepted')
 
-            filt_windows = []
-            for w in windows:
-                label_arr = label_source.get_labels(w).get_label_arr(w)
-                null_inds = (
-                    label_arr.ravel() == class_config.get_null_class_id())
-                if not np.all(null_inds):
-                    filt_windows.append(w)
-            windows = filt_windows
+        filt_windows = []
+        for w in windows:
+            chip = raster_source.get_chip(w)
+            nodata_below_thresh = nodata_below_threshold(
+                chip, chip_nodata_threshold, nodata_val=0)
+
+            label_arr = label_source.get_labels(w).get_label_arr(w)
+            null_labels = label_arr == class_config.get_null_class_id()
+
+            if not np.all(null_labels) and nodata_below_thresh:
+                filt_windows.append(w)
+        log.info('Label and NODATA filtering: '
+                 f'{len(filt_windows)}/{len(windows)} chips accepted')
+
+        windows = filt_windows
         return windows
 
-    def should_use_window(window):
+    def should_use_window(window: Box) -> bool:
         if co.negative_survival_prob >= 1.0:
             return True
         else:
@@ -80,18 +101,6 @@ def get_train_windows(scene, class_config, chip_size,
     return windows
 
 
-def fill_no_data(img, label_arr, null_class_id):
-    # If chip has null labels, fill in those pixels with
-    # nodata.
-    null_inds = label_arr.ravel() == null_class_id
-    img_shape = img.shape
-    if np.any(null_inds):
-        img = np.reshape(img, (-1, img_shape[2]))
-        img[null_inds, :] = 0
-        img = np.reshape(img, img_shape)
-    return img
-
-
 class SemanticSegmentation(RVPipeline):
     def __init__(self, config: 'RVPipelineConfig', tmp_dir: str):
         super().__init__(config, tmp_dir)
@@ -114,10 +123,17 @@ class SemanticSegmentation(RVPipeline):
             img_channels = scene.raster_source.num_channels
         return img_channels
 
+    def chip(self, *args, **kwargs):
+        log.info(f'Chip options: {self.config.chip_options}')
+        return super().chip(*args, **kwargs)
+
     def get_train_windows(self, scene):
-        return get_train_windows(scene, self.config.dataset.class_config,
-                                 self.config.train_chip_sz,
-                                 self.config.chip_options)
+        return get_train_windows(
+            scene,
+            self.config.dataset.class_config,
+            self.config.train_chip_sz,
+            self.config.chip_options,
+            chip_nodata_threshold=self.config.chip_nodata_threshold)
 
     def get_train_labels(self, window, scene):
         return scene.ground_truth_label_source.get_labels(window=window)